# app/Http/Controllers/ComicController.php
<?php

namespace App\Http\Controllers;

use App\Constants\Geral;
use App\Http\Requests\ComicRequest;
use App\Services\ComicService;
use Illuminate\Http\Request;

class ComicController extends Controller
{
    protected $service;

    public function __construct(ComicService $service)
    {
        $this->service = $service;
    }

    /**
     * @OA\Post(
     *      tags={"Comic"},
     *      path="/hq/cadastrar",
     *      @OA\Parameter(
     *          name="name",
     *          required=true,
     *      ),
     *      @OA\Parameter(
     *          name="publication_date",
     *          required=true,
     *      ),
     *      @OA\Parameter(
     *          name="type_comic_id",
     *          required=true,
     *      ),
     *      @OA\Parameter(
     *          name="user_id",
     *          required=true,
     *      ),
     *      security={{"bearerAuth": {}}},
     *      @OA\Response(response="200", description="Cadastra as informações da HQ"),
     *      @OA\Response(response="401", description="Usuário não Autenticado"),
     *      @OA\Response(response="422", description="Erro em algum campo obrigatório"),
     * )
     */
    public function create(ComicRequest $request)
    {
        $comic = $this->service->create($request);

        return ['status' => true, 'message' => Geral::HQ_CADASTRO, 'comic' => $comic];
    }

    /**
     * @OA\Get(
     *     tags={"Comic"},
     *     path="/hq/",
     *     security={{"bearerAuth": {}}},
     *     @OA\Response(response="200", description="Lista as HQ cadastradas por usuário"),
     *     @OA\Response(response="401", description="Usuário não Autenticado"),
     * )
     *
     * @param Request $request
     * @return App\Models\Comic
     */
    public function list(Request $request)
    {
        $comic = $this->service->list($request);

        return ['status' => true, 'message' => Geral::HQ_ENCONTRADOS, 'comic' => $comic];
    }

    /**
     * Display the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function show($id)
    {
        //
    }

    /**
     * Show the form for editing the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function edit($id)
    {
        //
    }

    /**
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function update(Request $request, $id)
    {
        //
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function destroy($id)
    {
        //
    }
}