# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gerenciamento de HQ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "parameters": [
                    {
                        "name": "email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário loga com email e senha"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Rota para deslogar usuário e revogar token"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hq/cadastrar": {
            "post": {
                "tags": [
                    "Comic"
                ],
                "operationId": "0f2c37ace41f88c5b9048a9319795668",
                "parameters": [
                    {
                        "name": "name",
                        "required": true
                    },
                    {
                        "name": "publication_date",
                        "required": true
                    },
                    {
                        "name": "type_comic_id",
                        "required": true
                    },
                    {
                        "name": "user_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra as informações da HQ"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hq/": {
            "get": {
                "tags": [
                    "Comic"
                ],
                "operationId": "9e02c0b19db5a5d55619fefee938cab4",
                "responses": {
                    "200": {
                        "description": "Lista as HQ cadastradas por usuário"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/editora/cadastrar": {
            "post": {
                "tags": [
                    "Publisher"
                ],
                "operationId": "4833aebcb40416fbdf59eb94f06229d4",
                "parameters": [
                    {
                        "name": "name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra as informações da Editora"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/editora/": {
            "get": {
                "tags": [
                    "Publisher"
                ],
                "operationId": "f17302ac8ad5cd16a69411e2052d0f64",
                "responses": {
                    "200": {
                        "description": "Lista as Edições cadastradas"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tipo_hq/cadastrar": {
            "post": {
                "tags": [
                    "TypeComic"
                ],
                "operationId": "474a8378c414d748b563a68449f57c49",
                "parameters": [
                    {
                        "name": "name",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra as informações do tipo de HQ"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tipo_hq/": {
            "get": {
                "tags": [
                    "TypeComic"
                ],
                "operationId": "ea456d629f7d5001da485698f7d70a72",
                "responses": {
                    "200": {
                        "description": "Lista os tipo de HQ cadastrados"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cadastrar": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "ec03a44772dc1a8e663f30e5fdcb9cd8",
                "parameters": [
                    {
                        "name": "name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cadastra as informações do usuário"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    },
                    "422": {
                        "description": "Erro em algum campo obrigatório"
                    }
                }
            }
        },
        "/user/": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "b334625968b998a2f5c9970e14bc21d1",
                "responses": {
                    "200": {
                        "description": "Apreseta informações do usuário logado"
                    },
                    "401": {
                        "description": "Usuário não Autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}